refactor: move splsah screen logic to a custom hook
